ci: skip endpoint requiring authentication
The `vsts/info` endpoint always requires authentication. In the .NET
version, this is handled by Scalar directly, but in the C version,
authentication is handled by the `gvfs-helper`.

Since this only really matters in CI (interactive usage of Scalar with
Azure Repos will ask the user to authenticate, but that's handled
conveniently by the credential helper), we added support to skip that
endpoint via the environment variable `SCALAR_TEST_SKIP_VSTS_INFO`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>