SparseSetTests: Ensure no edits remain before merge
Something strange happens with the MergeChangesOutOfTheConeWithConflict
when run in sequence with the other tests in this class. Some edited
files from previous tests cause confusion when running 'git merge'.

For some reason, this is causing a behavior change with the
sparse-index, but it seems that either behavior would be acceptable
here.

Another thing we could do is remove the `ShouldContain("Merge conflict")`
line and the test passes that way, too. This seems like a more robust
solution.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>