SparseSetTests: handle new present-despite-SKIP_WORKTREE behavior
In Git 2.36, including the change in af6a51875a (repo_read_index: clear
SKIP_WORKTREE bit from files present in worktree, 2022-01-14), the behavior
of sparse-checkout was changed to automatically force any file outside the
sparse-checkout that exists on-disk to not be sparse (i.e., remove
'SKIP_WORKTREE').

Because of this, the situation where adding a folder with a changed file on
disk created a conflict no longer exists - because the "conflicting" file is
already non-sparse, adding the folder to the sparse-checkout definition no
longer involves any potential overwrite.

Update the tests to accommodate this new behavior.

Signed-off-by: Victoria Dye <vdye@github.com>