DiagnoseTests: make failure more helpful
The pattern `haystack.Contains(needle).ShouldEqual(true)` will provide
totally unhelpful output in case of failure, as opposed to
`haystack.ShouldContain(needle)`. Let's use the latter.

This goes back all the way to 3c09517e (Current snapshot of GVFS,
2017-02-02), the initial commit.

This seems to have been the only instance of that pattern in the code.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>